move Movie class to test, so we can use it in unit tests
