+++ src/test/java/com/meilisearch/sdk/utils/Movie.java
@@ -1,4 +1,4 @@
-package com.meilisearch.sdk;
+package com.meilisearch.sdk.utils;
 
 public class Movie {
 	private long id;
